Rename "index" to "cursor" everywhere
The term "index" carries strong associations with numerical indices for arrays. Using the term "cursor" instead is intended to better convey their intended role as opaque types which are meaningful only to the sequences to which they pertain.

Of course, numerical indices can still be used as Flux cursors -- but as with pointers and iterators in the STL, it's better to use a different name for the generalised concept.