Fix missing FLUX_FWD in take_while()
This is it, this is the bug that has been causing our optimised builds to fail... we were unintentionally always taking an lvalue and constructing a `ref_adaptor`, i.e. a raw pointer, hence the problems when passing an rvalue.

This also revealed another bug -- we can't always form cursor_t<Base const> as a function argument, so make things templates to avoid it.