Simplify sort implementation a little
We'll avoid passing the projection around by bundling it up with the comparator before we begin.

This also adds a tiny hack to avoid a couple of calls to std::invoke in the 99% case where Proj=std::identity, which makes us a tiny bit faster than ranges::sort in debug mode.