Implement flux::to<template-name>
We try to use CTAD to work out what the deduced type name would be, and then call the other overload of `flux::to()` to do the actual construction.

If the container doesn't support CTAD, we fall back to trying to use the sequence value type as the (only) template arg for the container template -- but only when no other arguments are provided, to prevent allocator args from being silently ignored.