Implement recursive flux::to() calls
For example,

    flux::split(...).to<std::vector<std::string>>()

This was a bit trickier than I expected. It's possible to hit hard errors here because I don't think it's possible to do the required recursive concept checks to avoid it (or at least, I haven't yet worked out how to do it). But if my reading of the P1206 wording is correct, the same is true of ranges::to(), so we'll probably be all right.