Add simple_sequence_base
This is a sequence wrapper which allow you to define a sequence as a Rust/Swift/Flow style iterator, by writing a class which inherits from flux::simple_sequence_base and provides a member function named maybe_next() which returns something "optional-like" (for example, a std::optional, a boost::optional, or a raw pointer).

For example:

```cpp
struct some_ints : flux::simple_sequence_base<some_ints>
{
    int i = 0;

    std::optional<int> maybe_next()
    {
        if (i < 10) {
            return {i++};
        } else {
            return std::nullopt;
        }
    }
};
```

simple_sequence_base in turn inherits from lens_base and so provides all the usual member functions. It's for this reason that the require method is called maybe_next() rather than the more usual next() -- the latter would hide the declaration of flow_base::next, and so would require a using declaration to bring it back into scope.