Allow sequence impls to be nested classes
Previously (ignoring "simple sequences"), there were two ways of implementing the Sequence protocol:

 * Provide a specialisation of sequence_iface<MyType>
 * Add a bunch of member functions to MyType with the right names etc

The first case is necessary in order to add Sequence conformance to types we don't own and cannot modify. For types *do* control though, it's a bit of a pain, as the sequence implementation is separate from (and possibly quite distant from) the type to which it pertains.

The second approach of providing member functions is fine for tiny examples, but (at least until we get deducing this) isn't very useful for more complex sequences as we'd have to duplicate everything for const- and non-const fns; and it doesn't interact all that well with inheriting from lens_base, as we can end up hiding names that we would then need to bring back into scope.

This PR replaces the member function approach by instead allowing you to provide a nested class named `MyType::flux_sequence_iface` which has the same format as `sequence_iface<MyType>` -- that is, static methods and nested typedefs etc.

The idea is that this is the best of both worlds: just one way of writing things, but with two options of where to put it (either in-class, or separately).

This has the slight down-side that trivial examples become slightly more verbose, but I don't think any real-world code would actually suffer.