Trivially-copyable sequences can use pass by value
The whole idea behind this PR is that we don't want to implicitly take references when constructing adaptors. The simplest (and, in an ideal world, best) way of doing this would be for our adaptor wrapper functions to take their arguments by value, e.g.

    auto filter(sequence auto src, auto pred) -> sequence auto;

Whilst this is the ideal, it has a couple of downsides: it makes it much too easy to accidentally copy an expensive container, and it means that we have some extra moves that we don't really need.

Previously we avoided this by requiring the argument to be an rvalue, which prevents implicit copying and minimises the number of moves, but has the downside that it's a bit *too* restrictive: basically, it's annoying to have to write an explicit flux::copy(seq) (or auto(seq) in C++23) for cheap-to-copy types like std::reference_wrapper or flux::ref_adaptor.

So this commit changes things again: now, if an argument to an adaptor function is

 1. passed in as an lvalue, and
 2. trivially copyable, and
 3. its decayed type satisfies sequence

then we'll implicitly take a copy and instantiate the adaptor instance with that. Otherwise, you'll need to explicitly pass by refenece, or explicitly pass a copy, or explicitly move into the adaptor.

(I'm slightly tempted to further restrict this to "small" types, say sizeof(T) <= 2*sizeof(void*) to avoid implicitly copying e.g std::array<int, 10'000>, but we'll see how it goes for now.)