Add getlines sequence
Similar to range-v3, wraps std::getline() in the sequence protocol. Internally re-uses a std::string for each call to getline() to avoid allocations.

The element type of the sequence is std::string const&. Of all the options (prvalue string, string_view, non-const string& etc) it's the best default, as it allows users to make the choice of whether they want to take a copy of the buffer or not, but without many of the lifetime problems of string_view.