Add iota source
A sequence of integers, how hard could it be?

Well, naming is an issue for a start. "Iota" comes in for criticism but it's been part of the STL for 25 years, and is also used in D and Go, so we'll stick with it (and we could hardly use Python's "range")

As far as the implementation goes, this is a bit simpler than Ranges' `views::iota` -- we don't have a custom difference type any more, and limits are handled a bit differently.