Add compare() algorithm
This performs three-way comparison of the elements of a pair of sequences, equivalent to `std::lexicographical_compare_three_way` (which might be the worst function name in the whole stdlib). The result type must be one of the three standard library comparison categories.

At the moment, as with `equal()`, there is only a free function `flux::compare()`, because I dislike the asymmetry of `lhs.compare(rhs)`. I might change my mind on this in future, especially given `std::string[_view]`  has a precisely equivalent `compare()` member.