Add custom assertion functions
Depending on the selected error policy, this will either "panic" and throw an exception of type `flux::unrecoverable_error`, or cut out the middle man and just `std::terminate()` (optionally printing a message first). Hopefully the name "unrecoverable error" is plain enough that this means that we're in a bad state and can't continue.