Add from_range() source
We no longer want *all* ranges to automatically be sequences, only contiguous ones (because they can be cheaply indexed, which means easier bounds checking and less UB).

For compatibility reasons though, we do kind need to be able to interface with ranges code. So we'll add an adaptor which does that.

Because we want to enforce that cursor_t<Seq> and cursor_t<Seq const> are the same whenever Seq is const-iterable, we need to have decide up-front whether we want to use const_iterators for our sequence iteration. Of course, when views come into the picture we can have a very confusing situation where we say

    auto const seq = flux::from_crange(some_view);

...and yet we can still mutate the elements. Ah well.