Add [read|move]_at_unchecked() functions
read_at_unchecked() does the same as read_at(), but possibly avoids bounds checking. If a sequence doesn't specialise it, it just returns read_at().

move_at_unchecked() is similar, but first looks for a specialisation of move_at() before falling back to read_at_unchecked() (which might itself fall back to read_at()).

Also implementations to some, but not all, adaptors. We need to do a thorough update of everything anyway...