Add unchecked adaptor
This is a simple adaptor which acts like a passthrough except that it replaces each call to read_at() with a call to read_at_unchecked(), and similarly for move_at().

Obviously this makes it unsafe to use in general, but in cases where we are sure we won't accidentally read out of bounds it give improved performance.