Use unchecked() wrapper in sort()
So it turns out that our new "Bounds Check All The Things" policy didn't do great things for our sort benchmark. We were still faster than std::sort in most cases, but not by as much as before.

Fortunately, because we're confident in our sort implementation (I hope?), we should be fine to not do bounds checking internally.