Add flatten adaptor
This is our version of C++20's views::join

This commit actually has two completely separate implementations: the first is single-pass only while the second can go all the way to bidirectional + bounded if the underlying sequence supports it.

The multipass version is used when all of the following are true:
  * the outer sequence is multipass
  * the element type of the outer sequence is a reference type
  * the inner sequence is multipass

Otherwise, the single-pass version is used.