Fix GH Actions yet again
It's Monday, so that means GitHub Actions and/or Homebrew has changed something again.

It worked two days ago.

The Linux build scripts seem to be complaining that the binutils package isn't installed, which either means that the GCC packages have been updated so they don't need them anymore (undoing the changes in #11), or else it's some sort of mistake and binutils should be auto-installed but isn't.

We'll try the latter possibility first.