Add stride() impl for bidirectional sequences
For bidirectional sequences, stride is a bit more complicated, because we want to make sure that

    auto seq = flux::stride(std::vector{1, 2, 3, 4, 5}, 2).reverse();

yields the correct [5, 3, 1], rather than simply starting at the end and striding backwards (which would yield [4, 2]).

To do this, we use the same design as P1899, namely keeping track of the number of "missing" elements from within our cursor. This means that there is a bit of overhead when using stride() with bidir+ sequences, even if you don't actually want to iterate the resulting sequence backwards. Perhaps some way of limiting the capabilities of sequences (as we have in some of the tests) would be worth investigating.