Add multipass chunk() adaptor
This is the first of what will probably end up being three separate implementations of chunk_adaptor(). It handles multipass sequences, and the result is a multipass sequence.

This version is nice and simple: we can re-use the underlying cursor type, increment it by `chunk_sz` places on each call to `inc()`, and use `take(slice(), chunk_sz)` on each call to `read_at()`.

It is also a bounded_sequence when the underlying is bounded.