Avoid problematic constexpr test in MSVC
MSVC doesn't like constexpr evaluation of single-pass chunk -> flatten.

Specifically, the problem occurs in the first call to inc(seq, cur), complaining of a read of an uninitialised union member inside flatten's optional<inner_cur>. (The same happens if we use std::optional rather than flux::optional inside flatten_adaptor(), so it doesn't seem to be a bug in our optional implementation.)

The same check passes constexpr evaluation in GCC, and runtime eval with both GCC and MSVC (including GCC + UBSan), so I'm inclined to think it's a compiler bug rather than a Flux bug, although it is a little worrisome.