Add scan and exclusive_scan adaptors
These are the adaptor versions of std::inclusive_scan and std::exclusive_scan respectively, except that they don't have the same associativity requirements (so flux::scan() is more like std::partial_sum()).

They are single-pass only because they need to hold their state in the adaptor objects themselves.

(It might be possible to make them multipass by keeping the state in the cursor instead, but that makes me a bit nervous.)