Add slide adaptor
This takes a runtime window size `n`, and yields sliding windows of size `n` over the source sequence. For example, given a source sequence `[1, 2, 3, 4, 5]` and a window size of 2, will return

    [[1, 2], [2, 3], [3, 4], [4, 5]]

It is multipass only, since we need to use two cursors to keep track of the start and end of the window.