Add flux::proj and flux::proj2 function adaptors
So, we're doing away with projections. Which I'm a little bit sad about, because I actually like projections. But there is a non-zero cost in terms of compile times and complexity when it comes to specification and documentation, and I think there's a better way to offer the same functionality.

What we'll do instead is provide a pair function adaptors, `proj` and `proj2`.

The first one takes an n-ary function `F` and a unary projection function `P`. When invoked with a pack of arguments `a...`, `proj` calls `F(P(a)...)` -- that is, it applies the projection to each argument and calls `F` with the projected args.

`proj2` should really be called something like "binary_project" but that's too long... it does something  similar to the above but takes a binary function F and two unary projections P1 and P2 (either of which can be defaulted to std::identity). The result is a binary function object which when invoked with two args `a` and `b` calls `F(P1(a), P2(b))` -- that is, it applies the first projection to the first arg, and the second projection to the second arg.

The first function is more general, but the second is useful for things like `flux::equal(seq1, seq2, <comparator>)` where you want to apply a different projection to the elements of seq1 and seq2.