Rename proj::proj member variable
As a class with no user-declared constructors, it should be valid for `flux::proj` to have a non-static data member with the name `proj` -- and indeed both GCC and Clang accept it with no problem.

MSVC however doesn't seem to like it, giving error C2461 "constructor syntax missing formal parameters". This seems to be the same bug as https://developercommunity.visualstudio.com/t/generated-constructors-if-class-has-a-member-with/1626621, albeit with a different error message.

As a workaround we will -- with some regret -- rename the `proj::proj` member to `prj`