Add zip algorithms
This commit adds four new algorithms which duplicate the existing low-level single-sequence algorithms we already have, but operate on a variadic pack of sequences in lock-step.

First up is zip_for_each_while(), the multi-sequence analogue of our existing for_each_while() -- except that it's not a customisation point (though it will call for_each_while() if passed a single sequence). Like the single-sequence version, it's effectively find_if_not() with a weakened precondition on the passed-in predicate and forms the root of all of the other functions.

The other functions are:
 * `zip_for_each()`, which just calls `zip_for_each_while()` with a predicate that always returns true
 * `zip_find_if()` which just calls `zip_for_each_while()` with a negated predicate
 * `zip_fold()`, which calls `zip_for_each()` with a function that updates the accumulator in each call

So basically exactly what we do in the single-sequence case, but generalised to operate on multiple sequences at once.