Update drop() to handle out-of-range arguments
Previously, we assumed the argument to drop was in the range [0, size].

Now, we take any integral type as an argument and perform a checked cast to distance_t (incase the user has some somethign silly like pass ULONG_MAX). Then we check if the argument is less than zero, in which case we raise a runtime error.

After that, in the actual implementation of drop we handle oversized arguments by carefully making sure we never overstep the end of the sequence when dropping in first(). We also make sure that size() never returns a negative value, and data() returns the past-the-end pointer if the resulting sequence is empty.