Handle out-of-range arguments to take()
It turns out that unlike drop(), take() didn't have many problems with arguments that were too large, or negative. The only issue was in size(), which we could fix by clamping the result to zero.

Nonetheless,  for consistency with drop() (and to help the user diagnose errors) we'll raise a runtime error if we are passed a negative argument.