Add ends_with() algorithm
This one is a bit more complicated than starts_with().

We have a specialised version if both sequences are bidirectional and bounded, which is effectively `starts_with(reverse(seq1), reverse(seq2))` but doesn't actually use the reverse adaptor to avoid the usual extra cursor inc/dec in deref.

If the sequences are not both reversible then we first need to work out their lengths, and then fast-forward the haystack cursor to the right place and effectively do `equal()` from there. This requires that both sequences are either multipass or sized, because otherwise we'll "use up" single-pass sequences while finding their length.