Add scan_first adaptor
This is to `scan()` as `fold_first()` is to `fold()`: that is, the initial value is taken to be the first element of the sequence, and then we perform a scan over the remaining elements.

Like `fold_first()`, we need to use an optional (internal in this case) to handle the case where the sequence is empty.

This matches the behaviour of `std::inclusive_scan()` when not passed an initial value, whereas our `scan()` defaults the initial value to a default-initialised `value_t<Seq>`.