Add test macro for C++23 tuple common_reference
The paper (P2321) that introduced std::views::zip also added new converting constructors for tuple and pair, and specialisations of basic_common_reference for both types in order to make the std::common_reference_with requirements for C++20 iterators yielding tuples and pairs work correctly.

We've had the same requirements for sequences commented out since the very beginning, but now that zip is supported by both MSVC and GCC we can actually start to use them (at least, when compiled in C++23 mode with a supporting standard library).

As an added bonus -- or rather, the real reason I'm doing this now -- it means that zip(a, b).read_only() will return the correct type when a and b yield references.