Make sure we actually run tests at runtime
Because almost the entire library is constexpr, the majority of our tests look like this:

```cpp
constexpr bool test_xxx() {
    // test code
    return true;
}
static_assert(test_xxx());

TEST_CASE("xxx") {
    REQUIRE(test_xxx());
}
```

That is, we do a constexpr test during compilation (which has the added benefit of testing for many guises of UB), and then again at runtime using Catch.

Unfortunately, compilers are getting increasingly smart at noticing that they've already calculated the result of `test_xxx()`, so they can optimise the run-time version to just `REQUIRE(true)`, which isn't exactly in the spirit of what we want as it means we never actually execute the tests at runtime! It also means that the tested functions never show up in our code coverage measurements.

As a workaround, this commit adds an otherwise-useless `template <bool = true>` to every `test_xxx()` function. Then in the runtime version we explicitly call `test_xxx<false>()` to force a separate instantation which should hopefully show up in CodeCov.

This probably won't do wonders for our compile times, but we'll see...