Add cycle() adaptor
Given a multipass sequence, `cycle(seq)` is an `infinite_sequence` which endlessly yields the elements of `seq`, looping back to the beginning when the underlying sequence is exhausted.

To avoid "spooky action at a distance", `cycle()` only provides const access to the elements of the underlying sequence.

This commit also makes some changes to `take()` to properly handle infinite sequences. Other adaptors probably need similar treatment.