Add flux::cref()
Flux has had the `flux::ref()` function for a long time. This basically does exactly the same as `std::ref()`, except the argument is required to be an lvalue sequence, and  the result also provides the "inline sequence" member function API to allow chaining. Like `std::ref()`, calling `flux::ref()` on something that is already a ref_adaptor just returns a copy of the adaptor.

This commit adds `flux::cref()`, which unsurprisingly does the equivalent to `std::cref()` -- that is, it's the same as `flux::ref(std::as_const(seq))`. This requires that `Seq const` is a sequence.

`flux::cref()` should be preferred to `flux::ref()` pretty much everywhere unless you absolutely want a mutable reference to a sequence. As such, this commit also changes a bunch of uses of `ref()` into `cref()` in the tests.