Rename flux::ref to mut_ref and cref to ref
...and make flux::mut_ref only accept non-const lvalue args

In general, we want to avoid having more than one mutable reference to something at a time. But having `ref` be one character shorter than `cref` sort of implies that it's the default, in that it's what people will reach for first.

This commit reverses that, and so that `flux::ref()` take a const reference, and makes you say `flux::mut_ref(s)` to take a mutable reference to `s`. This is suitably long and ugly that it might make you question whether it' truly what you want. And yes, the name does look a bit "Rusty", but there are only so many ways to abbreviate the word "mutable"...