cursors() adaptor requires a multipass sequence
The previous constraint that the cursor type was copy-constructible isn't wrong as such, but there's no great way of using the cursors adaptor with a single-pass sequence, because what are you going to do with the returned cursor? The only way it could work would be to do something like `mut_ref(seq).cursors()` and then use the returned cursor with the original `seq`, but we want to discourage multiple mutable references like this.