Make from() accept adaptable_sequences only
This changes the from() function so that now wraps the passed-in sequence in a type that inherits from `inline_sequence_base`, allowing for chaining using the "dot"member syntax. If the passed-in sequence already inherits from `inline_sequence_base`, then it is returned by value unchanged.

This means that `from()` now rejects (non-trivially copyable) lvalues. The idea is to promote the use of either `ref()` (preferably) or `mut_ref()` (when necessary) to make it clear when references are being taken.

There is one case that this doesn't cover, which is when you have a forwarding reference to a sequence that you want to temporarily pass to an adaptor for the duration of a function call. We do this in two places in the codebase (namely in the implementations of `to()` and `sort()`, which will need to be changed.