Add flux::from_fwd_ref() function
This does what `from()` used to do: if passed an lvalue it will call `ref()` or `mut_ref()` (depending on whether the argument is const or not), and if passed an rvalue it will wrap it in an "inline sequence" to allow chaining.

We don't actually need the last bit in the two uses we have -- just passing through an rvalue would be fine -- but it probably does no harm.