Add flux::unfold()
`fold()` takes a sequence and a function and produces a single value; `unfold()` does the opposite -- it takes a single value and a function and produces a sequence.

This is basically the equivalent of the `successors()` function in Rust or `views::generate()` in Range-V3 (but without the race condition in the latter).

Thanks to Ben Deane for the name!