Disable unreachable code warning on MSVC
...when compiling tests.

We get lots of these when compiling with optimisation enabled, but they all seem to be false positives, e.g. complaining that the "else" branch of an if statement is unreachable, or the constructor member-init-list of `flux::optional`, which is really weird.

I guess a lot is happening in the front end due to all our tests being constexpr, so by the time the backend sees it there's a lot of dead code?