Add optional CMake module build
If CMake is configured with FLUX_BUILD_MODULE=On (default is Off), then we'll attempt to build a module using add_module_library() defined in modules.cmake.

We'll also attempt to build a small test program to verify that the module can be correctly imported.

Everything seems to work fine with Clang 16.

With GCC 13, the Ninja generator currently runs into problems* building the module. With a Makefile generator, building the module succeeds, but the import fails complaining it can't find the generated module. I don't know if this is caused by a GCC bug or by modules.cmake.

MSVC is not yet tested.

* https://gcc.gnu.org/bugzilla/show_bug.cgi?id=105467