Add adjacent_filter adaptor
`adjacent_filter()` takes a binary predicate and has the same semantics as Range-V3's views::adjacent_filter: elements are tested pair-wise, and if the predicate returns `false` then the second element of the pair is discarded.

This also adds `dedup()` as a shortcut for `adjacent_filter(std::not_equal_to{})`, because that's what people are probably going to want most of the time.

Fixes #116