Fix multidim benchmark CI failure
It looks like MSVC's STL now provides C++23 std::views::cartesian_product, which clashes with the polyfill version we use for the multidim benchmark.

Fortunately, there's a feature test macro we can use to detect whether the std version exists, so we'll use that to select the std version if possible or otherwise fall back to our version.