Make splitting more generic
Rather than a pattern, `split_adaptor` now takes a "Splitter", a function object which, when passed a sequence, returns the bounds of the next slice.

This opens the door to potentially doing more sophisticated splitting, for example using a (CTRE?) regex or one of the standard library searchers.

At the moment, we don't publicly expose the generic Splitter interface, but this could be done in future. Rather, our existing `split()` overloads have been changed to use new (internal) `pattern_splitter` and `delim_splitter` implementations.