Add optional<T&>::emplace
We've had optional<T>::emplace for a while. As you might expect, it constructs a T directly inside the optional, using the supplied arguments.

For the optional<T&> version, we require the argument to be a single argument of a "compatible" lvalue reference type -- that is, the same condition as the constructor.

This allows emplace to be used in generic code which doesn't know if it's dealing with optional<T> or <T&>.