More CMake improvements
This is an attempt to solve the CMake errors reported in #130. It uses the new target_sources(FILE_SET) command to set the sources for the `flux` target, which will hopefully avoid the problems we've been seeing. It also adds a `flux::flux` alias target, which apparently is a good thing, although I don't know enough CMake to understand why.

The FILE_SET usage means setting our cmake_minimum_required to 3.23, which is pretty new (March 2022) -- but still older than our oldest supported compiler (GCC 11.3, released April '22). I think the odds of someone using a cutting-edge compiler but an old cmake are pretty slim, so hopefully it won't be a problem.