array_ptr is implicitly constructible from ref()
Unlike std::span, our array_ptr is only explicitly constructible from contiguous sequences. But since flux::ref has pointer semantics, we should enable implicit conversion to array_ptr -- they're platonically the same thing (a pointer to a contiguous sequence) but the latter is type erased.

This means that given

     void my_function(flux::array_ptr<int const> span);

we can call it with

    std::vector vec = ...
    my_function(flux::ref(vec));

which is a bit nicer than before.

We could be implement this as either a conversion operator on flux::ref, or as a non-explicit constructor on flux::array_ptr; it turns out to be easier to do it as the latter, because we also need to add a new deduction guide if we want to be able to CTAD array_ptr(ref(seq)).