Roll back some reverse_adaptor constraint checks
Going "all in" with constraint checks works fine with GCC 12+ and Clang, but causes an ICE in GCC 11.3 and some sort of strange error in MSVC that I haven't worked out yet.

So let's roll back some of the changes -- while still retaining enough to fix the original bug -- and see what happens.