Attempt to work around Clang 18 ICE
This is a wild stab in the dark which attempts to solve the Clang 18 internal compiler error reported in #174.

Homebrew doesn't have LLVM 18 yet so I don't have a local install of the latest Clang, and we can't add it to CI either. This PR just slightly simplifies the code that Clang appears to dislike (based on the backtrace supplies in #174), but unfortunately I don't have any way of testing it yet.

So let's see what happens...