Specialise for_each_while for multipass + bounded
From #173,  it seems like the existing formulation of `for_each_while()` using an `is_last()` end check appears to get in the way of compiler auto-vectorisation in some circumstances.

For bounded + multipass sequences, we can instead save the end cursor as a local variable and perform an end check against that. This appears to make Clang more eager to vectorise the code in the loop, so let's do it.