Fix missing #include <algorithm> in zip.hpp
A couple of the zip functions use the init-list overload of std::min. Unfortunately this means we need to bring in the <algorithm> header (previously we happened to be getting away without it on all three platforms).

I'd rather avoid dragging in this large header, so a better long-term solution would be to add an overload of our own cmp::min which does the equivalent of std:min({ilist...})