Move tests from Catch to Doctest
Catch2 v3 is no longer a header-only library, and requires a build step. If it was just a single file taking a couple of seconds this would be no problem, but instead it's annoyingly slow -- building Catch takes far longer than actually building our own tests.

Even this might be worth it if we used all of Catch's advanced functionality, but really we barely scratch the surface of what it can do, just using the basic macros.

Instead of using Catch, this PR migrates to using [Doctest](https://github.com/doctest/doctest) instead, hopefully leading to faster build times.

Because Doctest uses (mostly) the same macro names as Catch, migration is pretty easy -- most of the diff is just changing headers. The only other difference is changing Catch's `SECTION(...)` to Doctest's `SUBCASE(...)`

It's worth noting that Doctest [may be unmaintained](https://github.com/doctest/doctest/issues/554). I don't consider this a huge risk at the moment because the library seems very stable, and it's well used within the C++ community. But if things start to bit-rot in future we can always move back to Catch, or some other test framework.