Add multipass flatten_with adaptor
As with flatten(), we have two separate implementations of flatten_with(), one of which is single-pass only and another which is multipass and potentially bidirectional.

The latter is used when the outer sequence, the inner sequence and the pattern are all multipass, and additionally the element type of the outer sequence is an lvalue reference (so we don't need any temporary storage).

It as additionally bidirectional when the outer sequence, the inner sequence and the pattern are all bidir, and the inner sequence and the pattern are both bounded.