Add .clang-format
After a lot of messing about, I've think I've finally come up with a clang-format config that I don't hate. It still messes up the concept definitions in `core/concepts.hpp`, but other that that it seems to do a pretty good job of matching the current layout.

I'm not planning on reformatting the whole codebase or anything like that but this should hopefully be useful for new contributors.

Supercedes #110