Remove simple_sequence
This was added very early in the life of Flux with the intention of making it easy to write Flow-style (or Rust-style) iterators with just a `next() -> optional` method.

It works fine, but it's not particularly useful -- if you're doing a quick prototype then `flux::generator` is quicker and easier, and otherwise a "full" sequence implementation isn't much more work and usually gets you multipass for free.

So given that it's not useful and it's not used anywhere, let's get rid of it.